Caching API Responses and the Requests that Made Them (#15)
Open-Meteo has a 10,000 soft limit that I'd like to adhere to. To do this, we can limit API requests so we only make one when we absolutely need new data. Part of this is local caching. This will not only cache the response in memory, but also store it to the device, so if it's dropped out of memory, we can still retain it. It'll also only update if the time has changed or the request itself is different. Hopefully, this can keep most people under very many API hits per day.