API Limiter (#16)
Creates an API limiter. 

In an effort to ensure we don't allow users to overload the server, on top of caching, we'll also have a hard limit on the number of API requests.

Since the data only changes once per hour, the only way to reasonably hit 20 requests per day would be to change location frequently or stay up for 20 hours, checking the weather every hour.

We may increase this to allow for settings and location changes, but, for now, and for testing, this will make sure we keep requests to a reasonable level and don't hammer Open Meteo.