Screen Options Optimized for Smaller Screens (#29)
Flip phones are making a comeback! The Moto Razr+ and Samsung Galazy Z Flip5 both feature big, gorgeous outer displays. Wouldn't it be cool if we had a version of the app that worked on those displays?
Boom!
This splits up the weather screen into two columns if, during recomposition, it detects that the screen is wider than it is tall, or close to that, anyway.
Here's the thing: Compose will recompose any time you close or open your foldable or flippable. Neat, right? A super simple way to adapt a composeable for a smaller display.

But wait, there's more!

We have null view states now, so we can load a loading screen by default and swap it out. Now we can use the same loading screen even for loading other views.

There's also an error screen.

We've also fixed the button and about button views on longer devices, and the user can pick a side, for left or right handed mode as well as using them with an outer display and those camera cutouts.